rename, reorganize; pass verbose prop
